Add observes property for add inventory command
Due to limitations in the Prism DelegeteCommand's observe function, only view model properties can be observed i.e. Person.Name is not allowed. And can only observe one property change to auto raise property changed for can execute to be called. So we use ObservesCanExecute on a boolean that is checked when we manually update NewItem instead